refactor(character): extract StatCell into its own file
